+++ packages/react/src/components/pagination/Pagination.test.tsx
@@ -14,6 +14,7 @@ const renderPagination = ({ pageCount, pageIndex, siblingCount }) => {
       pageHref={() => '#'}
       language="en"
       siblingCount={siblingCount}
+      dataTestId="hds-pagination"
     />,
   );